finished server side implementation regarding access rights

+++ lu.uni.lassy.icrash.desktop.dev.java.server/src/lu/uni/lassy/excalibur/examples/icrash/dev/java/system/IcrashSystemImpl.java
@@ -944,7 +944,7 @@ public class IcrashSystemImpl extends UnicastRemoteObject implements
 			if (currentRequestingAuthenticatedActor instanceof ActCoordinator) {
 				ActCoordinator theActCoordinator = (ActCoordinator) currentRequestingAuthenticatedActor;
 				//PreP3
-				if (checkAccessRights(theActCoordinator, aEtCrisisType)) {
+				if (checkAccessRights(theActCoordinator, theCrisis.type)) {
 					//PostF1
 					theCrisis.type = aEtCrisisType;
 					DbCrises.updateCrisis(theCrisis);
@@ -976,15 +976,16 @@ public class IcrashSystemImpl extends UnicastRemoteObject implements
 					.getValue());
 			if (currentRequestingAuthenticatedActor instanceof ActCoordinator) {
 				ActCoordinator theActCoordinator = (ActCoordinator) currentRequestingAuthenticatedActor;
-				//Check if the current requesting coordinator's access rights match the crises type
-				//if it does then the coordinator can continue if not he is not allowed to continue
-				//PostF1
-				theCrisis.status = aEtCrisisStatus;
-				DbCrises.updateCrisis(theCrisis);
-				PtString aMessage = new PtString("The crisis status has been updated !");
-				theActCoordinator.ieMessage(aMessage);
-	
-				return new PtBoolean(true);
+				//PreP3
+				if (checkAccessRights(theActCoordinator, theCrisis.type)) {
+					//PostF1
+					theCrisis.status = aEtCrisisStatus;
+					DbCrises.updateCrisis(theCrisis);
+					PtString aMessage = new PtString("The crisis status has been updated !");
+					theActCoordinator.ieMessage(aMessage);
+		
+					return new PtBoolean(true);
+				}
 			}
 		}
 		catch (Exception e){
@@ -1008,8 +1009,6 @@ public class IcrashSystemImpl extends UnicastRemoteObject implements
 			if (currentRequestingAuthenticatedActor instanceof ActCoordinator) {
 				ActCoordinator theActCoordinator = (ActCoordinator) currentRequestingAuthenticatedActor;
 				CtCoordinator theCtCoordinator = (CtCoordinator) getCtAuthenticated(theActCoordinator);
-				//Check if the current requesting coordinator's access rights match the crises type
-				//if it does then the coordinator can continue if not he is not allowed to continue
 				
 				log.debug("theCrisis Instance is " + theCrisis.toString());
 				log.debug("aDtCrisisID.value.getValue() is "
@@ -1022,46 +1021,48 @@ public class IcrashSystemImpl extends UnicastRemoteObject implements
 				else
 						log.debug("theCtCoordinatorAtPre is "
 								+ assCtCrisisCtCoordinator.get(theCrisis).toString());
-				//Check if the current requesting coordinator's access rights match the crises type
-				//if it does then the coordinator can continue if not he is not allowed to continue
-				//PostF1
-				theCrisis.status = EtCrisisStatus.handled;
-				DbCrises.updateCrisis(theCrisis);
-				
-				assCtCrisisCtCoordinator.put(theCrisis, theCtCoordinator);
-				DbCrises.bindCrisisCoordinator(theCrisis, theCtCoordinator);
-				PtString aMessage = new PtString(
-						"You are now considered as handling the crisis !");
-				theActCoordinator.ieMessage(aMessage);
-	
-				//PostF2
-				for (CtAlert theAlert : getAlertsByCrisis(theCrisis)) {
-					theAlert.isSentToCoordinator(theActCoordinator);
-				}
-				//PostF3
-				if (theCtCoordinatorAtPre != null) {
-					ActCoordinator theActCoordinatorAtPre = (ActCoordinator) assCtAuthenticatedActAuthenticated
-							.get(theCtCoordinatorAtPre);
-					log.info("One of the crisis you were handling is now handled by one of your colleagues!");
-					PtString aMessage2 = new PtString(
-							"One of the crisis you were handling is now handled by one of your colleagues!");
-					theActCoordinatorAtPre.ieMessage(aMessage2);
-				}
-	
-				//PostF4
-				for (CtAlert theAlert : getAlertsByCrisis(theCrisis))
-				{
-					Enumeration<CtAlert> enumKey = assCtAlertCtHuman.keys();
-					while(enumKey.hasMoreElements()){
-						CtAlert aCtAlert = enumKey.nextElement();
-						if (aCtAlert.equals(theAlert)){
-							CtHuman theHuman = assCtAlertCtHuman.get(aCtAlert);
-							if (!theHuman.isAcknowledged().getValue())
-								log.error("Unable to message a communication company about the crisis update");
+				//PreP3
+				if (checkAccessRights(theActCoordinator, theCrisis.type)) {
+					//PostF1
+					theCrisis.status = EtCrisisStatus.handled;
+					DbCrises.updateCrisis(theCrisis);
+					
+					assCtCrisisCtCoordinator.put(theCrisis, theCtCoordinator);
+					DbCrises.bindCrisisCoordinator(theCrisis, theCtCoordinator);
+					PtString aMessage = new PtString(
+							"You are now considered as handling the crisis !");
+					theActCoordinator.ieMessage(aMessage);
+		
+					//PostF2
+					for (CtAlert theAlert : getAlertsByCrisis(theCrisis)) {
+						theAlert.isSentToCoordinator(theActCoordinator);
+					}
+					//PostF3
+					if (theCtCoordinatorAtPre != null) {
+						ActCoordinator theActCoordinatorAtPre = (ActCoordinator) assCtAuthenticatedActAuthenticated
+								.get(theCtCoordinatorAtPre);
+						log.info("One of the crisis you were handling is now handled by one of your colleagues!");
+						PtString aMessage2 = new PtString(
+								"One of the crisis you were handling is now handled by one of your colleagues!");
+						theActCoordinatorAtPre.ieMessage(aMessage2);
+					}
+		
+					//PostF4
+					for (CtAlert theAlert : getAlertsByCrisis(theCrisis))
+					{
+						Enumeration<CtAlert> enumKey = assCtAlertCtHuman.keys();
+						while(enumKey.hasMoreElements()){
+							CtAlert aCtAlert = enumKey.nextElement();
+							if (aCtAlert.equals(theAlert)){
+								CtHuman theHuman = assCtAlertCtHuman.get(aCtAlert);
+								if (!theHuman.isAcknowledged().getValue())
+									log.error("Unable to message a communication company about the crisis update");
+							}
 						}
 					}
+					
+					return new PtBoolean(true);
 				}
-				return new PtBoolean(true);
 			}
 		}
 		catch (Exception e){
@@ -1085,19 +1086,20 @@ public class IcrashSystemImpl extends UnicastRemoteObject implements
 					.getValue());
 			if (currentRequestingAuthenticatedActor instanceof ActCoordinator) {
 				ActCoordinator theActCoordinator = (ActCoordinator) currentRequestingAuthenticatedActor;
-				//Check if the current requesting coordinator's access rights match the crises type
-				//if it does then the coordinator can continue if not he is not allowed to continue
-				//PostF1
-				theCrisis.comment = aDtComment;
-				DbCrises.updateCrisis(theCrisis);
-				PtString aMessage = new PtString("The crisis comment has been updated !");
-				try {
-					theActCoordinator.ieMessage(aMessage);
-				} catch (RemoteException e) {
-					Log4JUtils.getInstance().getLogger().error(e);
+				//Prep3
+				if (checkAccessRights(theActCoordinator, theCrisis.type)) {
+					//PostF1
+					theCrisis.comment = aDtComment;
+					DbCrises.updateCrisis(theCrisis);
+					PtString aMessage = new PtString("The crisis comment has been updated !");
+					try {
+						theActCoordinator.ieMessage(aMessage);
+					} catch (RemoteException e) {
+						Log4JUtils.getInstance().getLogger().error(e);
+					}
+		
+					return new PtBoolean(true);
 				}
-	
-				return new PtBoolean(true);
 			}
 		}
 		catch (Exception e){
@@ -1120,7 +1122,7 @@ public class IcrashSystemImpl extends UnicastRemoteObject implements
 				//go through all existing crises
 				for (String crisisKey : cmpSystemCtCrisis.keySet()) {
 					CtCrisis crisis = cmpSystemCtCrisis.get(crisisKey);
-					if (crisis.status.toString().equals(aEtCrisisStatus.toString()))
+					if (crisis.status.toString().equals(aEtCrisisStatus.toString()) && crisis.type.toString().equals("huge"))
 						//PostF1
 						crisis.isSentToCoordinator(aActCoordinator);
 				}
@@ -1178,38 +1180,39 @@ public class IcrashSystemImpl extends UnicastRemoteObject implements
 					.getValue());
 			if (currentRequestingAuthenticatedActor instanceof ActCoordinator) {
 				ActCoordinator theActCoordinator = (ActCoordinator) currentRequestingAuthenticatedActor;
-				//Check if the current requesting coordinator's access rights match the crises type
-				//if it does then the coordinator can continue if not he is not allowed to continue
-				//PostF1
-				theCrisis.status = EtCrisisStatus.closed;
-				DbCrises.updateCrisis(theCrisis);
-				//PostF2
-				assCtCrisisCtCoordinator.remove(theCrisis);
-				//PostF3
-				Collection<CtAlert> keys = assCtAlertCtCrisis.keySet();
-				CtAlert[] alertkeys = keys.toArray(new CtAlert[0]);
-				for (int i = 0; i < alertkeys.length; i++) {
-					CtAlert theAlert = alertkeys[i];
-					if (assCtAlertCtCrisis.get(theAlert) == theCrisis) {
-						DbAlerts.deleteAlert(theAlert);
-						assCtAlertCtCrisis.remove(theAlert);
-						cmpSystemCtAlert.remove(theAlert.id.value.getClass());
-						if (!assCtAlertCtCrisis.contains(theCrisis))
-							break;
+				//PreP3
+				if (checkAccessRights(theActCoordinator, theCrisis.type)) {
+					//PostF1
+					theCrisis.status = EtCrisisStatus.closed;
+					DbCrises.updateCrisis(theCrisis);
+					//PostF2
+					assCtCrisisCtCoordinator.remove(theCrisis);
+					//PostF3
+					Collection<CtAlert> keys = assCtAlertCtCrisis.keySet();
+					CtAlert[] alertkeys = keys.toArray(new CtAlert[0]);
+					for (int i = 0; i < alertkeys.length; i++) {
+						CtAlert theAlert = alertkeys[i];
+						if (assCtAlertCtCrisis.get(theAlert) == theCrisis) {
+							DbAlerts.deleteAlert(theAlert);
+							assCtAlertCtCrisis.remove(theAlert);
+							cmpSystemCtAlert.remove(theAlert.id.value.getClass());
+							if (!assCtAlertCtCrisis.contains(theCrisis))
+								break;
+						}
 					}
+		
+					//PostF4	
+					PtString aMessage = new PtString("The crisis "
+							//+ "with ID '"
+							//+ aDtCrisisID.value.getValue() + "' "
+									+ "is now closed !");
+					try {
+						theActCoordinator.ieMessage(aMessage);
+					} catch (RemoteException e) {
+						Log4JUtils.getInstance().getLogger().error(e);
+					}
+					return new PtBoolean(true);
 				}
-	
-				//PostF4	
-				PtString aMessage = new PtString("The crisis "
-						//+ "with ID '"
-						//+ aDtCrisisID.value.getValue() + "' "
-								+ "is now closed !");
-				try {
-					theActCoordinator.ieMessage(aMessage);
-				} catch (RemoteException e) {
-					Log4JUtils.getInstance().getLogger().error(e);
-				}
-				return new PtBoolean(true);
 			}
 		}
 		catch (Exception e){
@@ -1426,6 +1429,9 @@ public class IcrashSystemImpl extends UnicastRemoteObject implements
 		}
 	}
 	
+	/* (non-Javadoc)
+	 * @see lu.uni.lassy.excalibur.examples.icrash.dev.java.system.IcrashSystem#oeUpdateCoordinatorAccessRights(lu.uni.lassy.excalibur.examples.icrash.dev.java.system.types.primary.DtCoordinatorID)
+	 */
 	public PtBoolean oeUpdateCoordinatorAccessRights(DtCoordinatorID aDtCoordinatorID, EtCrisisType aAccessRights) throws java.rmi.RemoteException{
 		try {
 			//PreP1
